Final_Corporate_Actions/requirements.txt
streamlit>=1.28.0
langchain>=0.0.3
python-dotenv>=1.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
pillow>=10.0.0
python-dateutil>=2.8.2
requests>=2.31.0
azure-identity>=1.12.0